+++ src/main/java/janggi/config/AppConfig.java
@@ -1,13 +1,13 @@
 package janggi.config;
 
 import janggi.controller.JanggiController;
-import janggi.controller.JanggiService;
 import janggi.repository.game.GameRepository;
 import janggi.repository.game.JdbcGameRepository;
 import janggi.repository.movement.JdbcMovementRepository;
 import janggi.repository.movement.MovementRepository;
 import janggi.repository.piece.JdbcPieceRepository;
 import janggi.repository.piece.PieceRepository;
+import janggi.service.JanggiService;
 import janggi.view.InputView;
 import janggi.view.OutputView;
